Fix new Clang warnings for Tests

+++ Sources/Tests/testprotocolmso2.h
@@ -45,8 +45,8 @@ namespace UnitTests {
     private:
         static const int SHORT_SLEEP;
         static const int LONG_TIMEOUT;
-        static const int BASE_PORT;
-        static int PORT_CNT;
+        static const uint16_t BASE_PORT;
+        static uint16_t PORT_CNT;
     };
 
 }
