+++ Sources/Tests/testtcpserver.h
@@ -23,8 +23,8 @@ namespace UnitTests
         static void Start_Restart_With_Client();
     private:
         static const int LONG_TIMEOUT;
-        static const int BASE_PORT;
-        static int PORT_CNT;
+        static const uint16_t BASE_PORT;
+        static uint16_t PORT_CNT;
     };
 
 }
